refactor(agent/surfaces): move agent/web to agent/surfaces/web
Pure relocation for epic 1193. Module path
github.com/panyam/mcpkit/agent/web -> .../agent/surfaces/web;
every relative replace directive recomputed one level deeper.
cmd/agentweb stays nested. No behavior change.
