+++ agent/surfaces/web/gen/go/mcpkit/agentweb/v1/agentwebv1connect/host.connect.go
@@ -8,7 +8,7 @@ import (
 	connect "connectrpc.com/connect"
 	context "context"
 	errors "errors"
-	v1 "github.com/panyam/mcpkit/agent/web/gen/go/mcpkit/agentweb/v1"
+	v1 "github.com/panyam/mcpkit/agent/surfaces/web/gen/go/mcpkit/agentweb/v1"
 	http "net/http"
 	strings "strings"
 )
