+++ agent/surfaces/web/go.mod
@@ -1,4 +1,4 @@
-module github.com/panyam/mcpkit/agent/web
+module github.com/panyam/mcpkit/agent/surfaces/web
 
 go 1.26.5
 
@@ -48,22 +48,22 @@ require (
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
-replace github.com/panyam/mcpkit => ../..
+replace github.com/panyam/mcpkit => ../../..
 
-replace github.com/panyam/mcpkit/agent => ../
+replace github.com/panyam/mcpkit/agent => ../../
 
-replace github.com/panyam/mcpkit/agent/host => ../host
+replace github.com/panyam/mcpkit/agent/host => ../../host
 
-replace github.com/panyam/mcpkit/ext/auth => ../../ext/auth
+replace github.com/panyam/mcpkit/ext/auth => ../../../ext/auth
 
-replace github.com/panyam/mcpkit/ext/skills => ../../ext/skills
+replace github.com/panyam/mcpkit/ext/skills => ../../../ext/skills
 
-replace github.com/panyam/mcpkit/ext/tasks => ../../ext/tasks
+replace github.com/panyam/mcpkit/ext/tasks => ../../../ext/tasks
 
-replace github.com/panyam/mcpkit/experimental/ext/agents => ../../experimental/ext/agents
+replace github.com/panyam/mcpkit/experimental/ext/agents => ../../../experimental/ext/agents
 
-replace github.com/panyam/mcpkit/experimental/ext/agents/clients/go => ../../experimental/ext/agents/clients/go
+replace github.com/panyam/mcpkit/experimental/ext/agents/clients/go => ../../../experimental/ext/agents/clients/go
 
-replace github.com/panyam/mcpkit/experimental/ext/events => ../../experimental/ext/events
+replace github.com/panyam/mcpkit/experimental/ext/events => ../../../experimental/ext/events
 
-replace github.com/panyam/mcpkit/experimental/ext/events/clients/go => ../../experimental/ext/events/clients/go
+replace github.com/panyam/mcpkit/experimental/ext/events/clients/go => ../../../experimental/ext/events/clients/go
