+++ agent/surfaces/web/sessions_test.go
@@ -15,8 +15,8 @@ import (
 
 	"github.com/panyam/mcpkit/agent"
 	"github.com/panyam/mcpkit/agent/host"
-	agentwebv1 "github.com/panyam/mcpkit/agent/web/gen/go/mcpkit/agentweb/v1"
-	"github.com/panyam/mcpkit/agent/web/gen/go/mcpkit/agentweb/v1/agentwebv1connect"
+	agentwebv1 "github.com/panyam/mcpkit/agent/surfaces/web/gen/go/mcpkit/agentweb/v1"
+	"github.com/panyam/mcpkit/agent/surfaces/web/gen/go/mcpkit/agentweb/v1/agentwebv1connect"
 	"github.com/panyam/mcpkit/core"
 )
