+++ agent/surfaces/web/shell.go
