+++ agent/surfaces/web/web/src/MemoryPanel.tsx
