+++ agent/surfaces/web/web/src/ToolsPanel.tsx
