+++ agent/surfaces/web/web/src/gen/mcpkit/agentweb/v1/host_pb.ts
