+++ agent/surfaces/web/web/src/panels.ts
